+++ instancing/instancing.cpp
@@ -95,7 +95,7 @@ public:
 	{
 		title = "Vulkan Example - Instanced mesh rendering";
 		enableTextOverlay = true;
-		srand(time(NULL));
+		srand((unsigned int)time(NULL));
 		zoom = -18.5f;
 		rotation = { -17.2f, -4.7f, 0.0f };
 		cameraPos = { 5.5f, -1.85f, 0.0f };
@@ -200,7 +200,7 @@ public:
 
 		VkDescriptorPoolCreateInfo descriptorPoolInfo =
 			vks::initializers::descriptorPoolCreateInfo(
-				poolSizes.size(),
+				static_cast<uint32_t>(poolSizes.size()),
 				poolSizes.data(),
 				2);
 
@@ -226,7 +226,7 @@ public:
 		VkDescriptorSetLayoutCreateInfo descriptorLayout =
 			vks::initializers::descriptorSetLayoutCreateInfo(
 				setLayoutBindings.data(),
-				setLayoutBindings.size());
+				static_cast<uint32_t>(setLayoutBindings.size()));
 
 		VK_CHECK_RESULT(vkCreateDescriptorSetLayout(device, &descriptorLayout, nullptr, &descriptorSetLayout));
 
@@ -309,7 +309,7 @@ public:
 		VkPipelineDynamicStateCreateInfo dynamicState =
 			vks::initializers::pipelineDynamicStateCreateInfo(
 				dynamicStateEnables.data(),
-				dynamicStateEnables.size(),
+				static_cast<uint32_t>(dynamicStateEnables.size()),
 				0);
 
 		// Load shaders
@@ -399,7 +399,7 @@ public:
 
 	float rnd(float range)
 	{
-		return range * (rand() / double(RAND_MAX));
+		return float(range * (rand() / double(RAND_MAX)));
 	}
 
 	void prepareInstanceData()
@@ -407,7 +407,7 @@ public:
 		std::vector<InstanceData> instanceData;
 		instanceData.resize(INSTANCE_COUNT);
 
-		std::mt19937 rndGenerator(time(NULL));
+		std::mt19937 rndGenerator((unsigned int)time(NULL));
 		std::uniform_real_distribution<float> uniformDist(0.0, 1.0);
 
 		// Distribute rocks randomly on two different rings
@@ -420,20 +420,20 @@ public:
 
 			// Inner ring
 			rho = sqrt((pow(ring0[1], 2.0f) - pow(ring0[0], 2.0f)) * uniformDist(rndGenerator) + pow(ring0[0], 2.0f));
-			theta = 2.0 * M_PI * uniformDist(rndGenerator);
+			theta = (float)(2.0 * M_PI * uniformDist(rndGenerator));
 			instanceData[i].pos = glm::vec3(rho*cos(theta), uniformDist(rndGenerator) * 0.5f - 0.25f, rho*sin(theta));
 			instanceData[i].rot = glm::vec3(M_PI * uniformDist(rndGenerator), M_PI * uniformDist(rndGenerator), M_PI * uniformDist(rndGenerator));
 			instanceData[i].scale = 1.5f + uniformDist(rndGenerator) - uniformDist(rndGenerator);
-			instanceData[i].texIndex = rnd(textures.rocks.layerCount);
+			instanceData[i].texIndex = (uint32_t)rnd((float)textures.rocks.layerCount);
 			instanceData[i].scale *= 0.75f;
 
 			// Outer ring
 			rho = sqrt((pow(ring1[1], 2.0f) - pow(ring1[0], 2.0f)) * uniformDist(rndGenerator) + pow(ring1[0], 2.0f));
-			theta = 2.0 * M_PI * uniformDist(rndGenerator);
+			theta = (float)(2.0 * M_PI * uniformDist(rndGenerator));
 			instanceData[i + INSTANCE_COUNT / 2].pos = glm::vec3(rho*cos(theta), uniformDist(rndGenerator) * 0.5f - 0.25f, rho*sin(theta));
 			instanceData[i + INSTANCE_COUNT / 2].rot = glm::vec3(M_PI * uniformDist(rndGenerator), M_PI * uniformDist(rndGenerator), M_PI * uniformDist(rndGenerator));
 			instanceData[i + INSTANCE_COUNT / 2].scale = 1.5f + uniformDist(rndGenerator) - uniformDist(rndGenerator);
-			instanceData[i + INSTANCE_COUNT / 2].texIndex = rnd(textures.rocks.layerCount);
+			instanceData[i + INSTANCE_COUNT / 2].texIndex = (uint32_t)rnd((float)textures.rocks.layerCount);
 			instanceData[i + INSTANCE_COUNT / 2].scale *= 0.75f;
 		}
